Implemented saving task to the database; Created a universal GameCodeManagement for handling states.

+++ tgbot/handlers/admin_panel/game_codes_handlers.py
@@ -3,12 +3,14 @@ from aiogram.fsm.context import FSMContext
 from aiogram.types import CallbackQuery, Message
 from aiogram.utils.i18n import gettext as _
 
+from infrastructure.repositories.game_task_repo import GameTaskRepository
 from tgbot.keyboards.admin_panel.game_codes_kb import (
     get_admin_panel_codes_kb,
     get_cancel_game_code_action_kb,
     get_game_codes_actions_kb,
 )
-from tgbot.states.game_code_state import AddGameCode
+from tgbot.services.game_task_service import GameTaskService
+from tgbot.states.game_code_state import GameCodeManagement
 
 router = Router()
 
@@ -26,59 +28,80 @@ async def manage_codes_handler(callback_query: CallbackQuery) -> None:
 @router.callback_query(F.data.startswith('admin_codes_for_'))
 async def actions_with_game_codes_handler(callback_query: CallbackQuery, state: FSMContext) -> None:
     game_name: str = callback_query.data.split('_')[-1]
-    await state.update_data(GameName=game_name)
+    await state.update_data(selected_game=game_name)
+    await state.set_state(GameCodeManagement.WaitingForActionSelection)
     await callback_query.answer()
     await callback_query.message.delete()
     await callback_query.message.answer(
-        text=_('<b>Select an action:</b>'),
+        text=_('You have selected a game: <b>{game_name}</b>\n\nSelect an action:').format(game_name=game_name),
         reply_markup=get_game_codes_actions_kb()
     )
 
 
 @router.callback_query(F.data == 'add_code')
 async def add_game_code_handler(callback_query: CallbackQuery, state: FSMContext) -> None:
-    await state.set_state(AddGameCode.Question)
+    data = await state.get_data()
+    if 'selected_game' not in data:
+        await callback_query.answer(
+            text=_('⚠️ First, choose a game!'),
+            show_alert=True)
+        return
+    await state.set_state(GameCodeManagement.WaitingForTask)
     await callback_query.answer()
     await callback_query.message.delete()
     await callback_query.message.answer(
-        text=_('🐥 Enter a question for the game:'),
+        text=_('🐥 Enter the task for the selected game:'),
         reply_markup=get_cancel_game_code_action_kb()
     )
 
 
-@router.message(AddGameCode.Question)
-async def process_add_game_task_handler(message: Message, state: FSMContext) -> None:
-    if len(message.text) < 1:
+@router.message(GameCodeManagement.WaitingForTask)
+async def process_task_input_handler(message: Message, state: FSMContext) -> None:
+    task = message.text
+    if len(task) < 1:
         await message.answer(
-            text=_('⚠️ The length of a question must be longer than one character!'),
+            text=_('⚠️ The length of a task must be longer than one character!'),
             reply_markup=get_cancel_game_code_action_kb()
         )
         return
-    await state.update_data(Question=message.text)
-    await state.set_state(AddGameCode.Answer)
+    await state.update_data(task=task)
+    await state.set_state(GameCodeManagement.WaitingForAnswer)
     await message.answer(
         text=_('⛱️ Enter an answer for the game:'),
         reply_markup=get_cancel_game_code_action_kb()
     )
 
 
-@router.message(AddGameCode.Answer)
-async def process_add_game_answer_handler(message: Message, state: FSMContext) -> None:
-    if len(message.text) < 1:
+@router.message(GameCodeManagement.WaitingForAnswer)
+async def process_answer_input_handler(message: Message, state: FSMContext, game_task_repo: GameTaskRepository) -> None:
+    answer = message.text
+    if len(answer) < 1:
         await message.answer(
             text=_('⚠️ The length of a answer must be longer than one character!'),
             reply_markup=get_cancel_game_code_action_kb()
         )
         return
+
     data = await state.get_data()
-    game_name: str = data['GameName']
-    question: str = data['Question']
-    answer: str = message.text
+    game_name: str = data['selected_game']
+    task = data['task']
+
+    new_code = await GameTaskService.create_task(
+        game_name=game_name,
+        task=task,
+        answer=answer,
+        game_task_repo=game_task_repo
+    )
     await state.clear()
     await message.answer(
-        text=_('{game_name}, {question}, {answer}'
-               '\n'
-               'Select a game to add a new code or go back to the main menu:').format(game_name=game_name, question=question, answer=answer),
+        text=_('✅ <i>Code for the game <b>{game}</b> successfully added!</i>\n\n'
+               '🐥 <i>Task: <b>{task}</b></i>\n'
+               '⛱️ <i>Answer: <b>{answer}</b></i>\n\n'
+               'Select a game to add a new code or go back to the main menu:').format(
+            game=new_code.game_name,
+            task=new_code.task,
+            answer=new_code.answer
+        ),
         reply_markup=get_admin_panel_codes_kb()
     )
